Refactored file structures
