+++ src/main/java/duke/Duke.java
@@ -8,10 +8,10 @@ import duke.commands.CommandManager;
 import duke.exceptions.DukeException;
 import duke.util.Ui;
 import duke.models.patients.PatientManager;
-import duke.models.assignedPatientTasks.AssignedTaskManager;
+import duke.models.assignedtasks.AssignedTaskManager;
 import duke.models.counter.Counter;
 import duke.storages.StorageManager;
-import duke.models.patientTasks.TaskManager;
+import duke.models.tasks.TaskManager;
 
 /**
  * Represents Duke, a Personal Assistant to help
