+++ src/main/java/duke/commands/assignedtask/FindAssignedTaskCommand.java
@@ -1,18 +1,19 @@
-package duke.commands;
+package duke.commands.assignedtask;
 
+import duke.commands.Command;
 import duke.exceptions.DukeException;
 import duke.util.Ui;
 import duke.models.patients.Patient;
 import duke.models.patients.PatientManager;
-import duke.models.assignedPatientTasks.AssignedTask;
-import duke.models.assignedPatientTasks.AssignedTaskManager;
+import duke.models.assignedtasks.AssignedTask;
+import duke.models.assignedtasks.AssignedTaskManager;
 import duke.storages.StorageManager;
-import duke.models.patientTasks.Task;
-import duke.models.patientTasks.TaskManager;
+import duke.models.tasks.Task;
+import duke.models.tasks.TaskManager;
 
 import java.util.ArrayList;
 
-public class FindPatientTaskCommand implements Command {
+public class FindAssignedTaskCommand implements Command {
 
     private String command;
 
@@ -21,7 +22,7 @@ public class FindPatientTaskCommand implements Command {
      *
      * @param cmd .
      */
-    public FindPatientTaskCommand(String cmd) {
+    public FindAssignedTaskCommand(String cmd) {
         super();
         this.command = cmd;
     }
