+++ src/main/java/duke/commands/functional/DukeCommand.java
@@ -1,13 +1,17 @@
-package duke.commands;
+package duke.commands.functional;
 
+import duke.commands.Command;
 import duke.exceptions.DukeException;
+import duke.util.ShortCutter;
 import duke.util.Ui;
 import duke.models.patients.PatientManager;
-import duke.models.assignedPatientTasks.AssignedTaskManager;
+import duke.models.assignedtasks.AssignedTaskManager;
+import duke.models.counter.Counter;
 import duke.storages.StorageManager;
-import duke.models.patientTasks.TaskManager;
+import duke.models.tasks.TaskManager;
 
-public class UndoCommand implements Command {
+
+public class DukeCommand implements Command {
 
     /**
      * .
@@ -22,10 +26,11 @@ public class UndoCommand implements Command {
     @Override
     public void execute(AssignedTaskManager patientTask, TaskManager tasks, PatientManager patientManager,
                         Ui ui, StorageManager storageManager) throws DukeException {
-        storageManager.savePatients(patientManager.getPatientList());
-        storageManager.saveTasks(tasks.getTaskList());
-        storageManager.saveAssignedTasks(patientTask.fullPatientTaskList());
-        ui.showUndoSuccess();
+
+        Counter counter = new Counter(storageManager.loadCommandFrequency());
+        ShortCutter shortCutter = new ShortCutter(counter, ui);
+        shortCutter.runShortCut().execute(patientTask, tasks, patientManager, ui, storageManager);
+
     }
 
     /**
