+++ src/main/java/duke/commands/functional/ExitCommand.java
@@ -1,11 +1,12 @@
-package duke.commands;
+package duke.commands.functional;
 
+import duke.commands.Command;
 import duke.exceptions.DukeException;
 import duke.util.Ui;
 import duke.models.patients.PatientManager;
-import duke.models.assignedPatientTasks.AssignedTaskManager;
+import duke.models.assignedtasks.AssignedTaskManager;
 import duke.storages.StorageManager;
-import duke.models.patientTasks.TaskManager;
+import duke.models.tasks.TaskManager;
 
 /**
  * Represents a command to exit Duke. The command.ExitCommand class
