+++ src/main/java/duke/commands/task/AddTaskCommand.java
@@ -1,14 +1,15 @@
-package duke.commands;
+package duke.commands.task;
 
+import duke.commands.Command;
 import duke.exceptions.DukeException;
 import duke.util.Ui;
 import duke.models.patients.PatientManager;
-import duke.models.assignedPatientTasks.AssignedTaskManager;
+import duke.models.assignedtasks.AssignedTaskManager;
 import duke.storages.StorageManager;
-import duke.models.patientTasks.Task;
-import duke.models.patientTasks.TaskManager;
+import duke.models.tasks.Task;
+import duke.models.tasks.TaskManager;
 
-public class AddStandardTaskCommand implements Command {
+public class AddTaskCommand implements Command {
     private Task newStandardTask;
 
     /**
@@ -16,7 +17,7 @@ public class AddStandardTaskCommand implements Command {
      *
      * @param taskDescription .
      */
-    public AddStandardTaskCommand(String taskDescription) {
+    public AddTaskCommand(String taskDescription) {
         super();
         this.newStandardTask = new Task(taskDescription);
