+++ src/main/java/duke/commands/task/DeleteTaskCommand.java
@@ -1,16 +1,17 @@
 //@@author kkeejjuunn
 
-package duke.commands;
+package duke.commands.task;
 
+import duke.commands.Command;
 import duke.exceptions.DukeException;
 import duke.util.Ui;
 import duke.models.patients.Patient;
 import duke.models.patients.PatientManager;
-import duke.models.assignedPatientTasks.AssignedTask;
-import duke.models.assignedPatientTasks.AssignedTaskManager;
+import duke.models.assignedtasks.AssignedTask;
+import duke.models.assignedtasks.AssignedTaskManager;
 import duke.storages.StorageManager;
-import duke.models.patientTasks.Task;
-import duke.models.patientTasks.TaskManager;
+import duke.models.tasks.Task;
+import duke.models.tasks.TaskManager;
 
 import java.util.ArrayList;
