+++ src/main/java/duke/models/tasks/Task.java
@@ -1,4 +1,4 @@
-package duke.models.patientTasks;
+package duke.models.tasks;
 
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleStringProperty;
