+++ src/main/java/duke/util/mementopattern/MementoParser.java
@@ -1,15 +1,15 @@
 package duke.util.mementopattern;
 
-import duke.commands.AddPatientCommand;
-import duke.commands.AddStandardTaskCommand;
-import duke.commands.AssignTaskToPatientCommand;
+import duke.commands.patient.AddPatientCommand;
+import duke.commands.task.AddTaskCommand;
+import duke.commands.assignedtask.AssignTaskToPatientCommand;
 import duke.commands.Command;
-import duke.commands.DeletePatientCommand;
-import duke.commands.DeletePatientTaskCommand;
-import duke.commands.DeleteTaskCommand;
-import duke.commands.UpdatePatientCommand;
-import duke.commands.UpdateTaskCommand;
-import duke.commands.UndoCommand;
+import duke.commands.patient.DeletePatientCommand;
+import duke.commands.assignedtask.DeleteAssignedTaskCommand;
+import duke.commands.task.DeleteTaskCommand;
+import duke.commands.patient.UpdatePatientCommand;
+import duke.commands.task.UpdateTaskCommand;
+import duke.commands.functional.UndoCommand;
 
 public class MementoParser {
 
@@ -18,9 +18,9 @@ public class MementoParser {
      * @return .
      */
     public static String getSaveFlag(Command command) {
-        if ((command instanceof AddPatientCommand) || (command instanceof AddStandardTaskCommand)
+        if ((command instanceof AddPatientCommand) || (command instanceof AddTaskCommand)
                 || (command instanceof AssignTaskToPatientCommand) || (command instanceof DeletePatientCommand)
-                || (command instanceof DeletePatientTaskCommand) || (command instanceof DeleteTaskCommand)
+                || (command instanceof DeleteAssignedTaskCommand) || (command instanceof DeleteTaskCommand)
                 || (command instanceof UpdatePatientCommand) || (command instanceof UpdateTaskCommand)) {
             return "save";
         } else if (command instanceof UndoCommand) {
